r.terraflow: allow larger maps  (#265)
* allow larger maps; patch provided by @metzm
* add testsuite

See: https://trac.osgeo.org/grass/ticket/1421

+++ raster/r.terraflow/types.h
@@ -26,8 +26,8 @@
 /* input parameters type */
 /* ------------------------------------------------------------ */
 
-typedef short dimension_type; /* represent dimension of the grid */
-static const dimension_type dimension_type_max=SHRT_MAX;
+typedef int dimension_type; /* represent dimension of the grid */
+static const dimension_type dimension_type_max=INT_MAX;
 
 typedef short direction_type;  /* represent the direction of a cell */
 static const direction_type DIRECTION_UNDEF=-1;
